Refactored check for water meter accessability
The problem with determining the IP address of IOT devices is more
widespread than this app, so the search algorithm has been factored out
into a separate module.

zoneNumber has been changed from a string to an int as there were some
cases trying to use the value as a string.

the valve_state class has been renamed zone_state, and setting the
setup of the values used in setting up the zones dictonary have been
changed to use 'zone' rather than 'valve' for consistancy.

The dictonary returned by the get_zones() function is now indexed by
the integer value of the zoneNumber.

Access to the Rachio API now will retry if the access times out